refactor: removed old datatable
