+++ package.json
@@ -29,7 +29,6 @@
     "@supabase/supabase-js": "^2.45.6",
     "@tanstack/react-query": "^5.59.16",
     "@tanstack/react-query-devtools": "^5.56.2",
-    "@tanstack/react-table": "^8.20.5",
     "@vercel/analytics": "^1.3.1",
     "axios": "^1.7.7",
     "class-variance-authority": "^0.7.0",
@@ -43,10 +42,8 @@
     "nuqs": "^1.19.3",
     "react": "19.0.0-rc-69d4b800-20241021",
     "react-dom": "19.0.0-rc-69d4b800-20241021",
-    "recharts": "^2.13.0",
     "tailwind-merge": "^2.5.2",
-    "tailwind-scrollbar": "^3.1.0",
-    "undici": "^6.19.8"
+    "tailwind-scrollbar": "^3.1.0"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.15",
